Various Resharper tidy ups
