+++ src/Crichton.Representors/Crichton.Representors.csproj
@@ -34,6 +34,8 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="CrichtonRepresentor.cs" />
+    <Compile Include="CrichtonTransition.cs" />
+    <Compile Include="CrichtonTransitionAttribute.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Serializers\HalSerializer.cs" />
   </ItemGroup>
