+++ src/Crichton.Representors/CrichtonTransition.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+
+namespace Crichton.Representors
+{
+    public class CrichtonTransition
+    {
+        public string Rel { get; set; }
+        public string Uri { get; set; }
+        public string TemplatedUri { get; set; }
+        public string InterfaceMethod { get; set; }
+        public IList<CrichtonTransitionAttribute> Attributes { get; set; } 
+    }
+}
