+++ tests/Crichton.Representors.Tests/CrichtonRepresentorTests.cs
@@ -1,13 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using FluentAssertions;
+﻿using FluentAssertions;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
 using Ploeh.AutoFixture;
-using Rhino.Mocks;
 
 namespace Crichton.Representors.Tests
 {
