+++ tests/Crichton.Representors.Tests/TestWithFixture.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Ploeh.AutoFixture;
+﻿using Ploeh.AutoFixture;
 using Ploeh.AutoFixture.AutoRhinoMock;
 
 namespace Crichton.Representors.Tests
